+++ CHANGELOG.md
@@ -8,6 +8,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 For compatibility information see `govukFrontendVersion` and `hmrcFrontendVersion` in
 [LibDependencies](project/LibDependencies.scala)
 
+## [3.25.0] - 2022-09-12
+
+### Changed
+
+- Updated `hmrc-frontend` to v5.7.0
+- Add `data-synchronise-tabs` to `HmrcTimeoutDialog` and `synchroniseTabs` to `TimeoutDialog.scala`
+- Add `synchroniseTabs` to `HmrcTimeoutDialogHelper` and use `reference.conf`
+
+### Compatible with
+
+- [hmrc/hmrc-frontend v5.7.0](https://github.com/hmrc/hmrc-frontend/releases/tag/v5.7.0)
+- [alphagov/govuk-frontend v4.3.1](https://github.com/alphagov/govuk-frontend/releases/tag/v4.3.1)
 
 ## [3.24.0] - 2022-08-31